chore : (PR 이슈 반영) 이미지 업로더 코드 중복 해결, 컴포넌트에서 서비스로 변경, 로컬 업로더 이미지 도메인으로 이동
