+++ backend/src/main/java/com/hjjang/backend/domain/image/service/ImageService.java
@@ -1,19 +1,41 @@
 package com.hjjang.backend.domain.image.service;
 
+import com.hjjang.backend.domain.image.domain.repository.ImageRepository;
 import com.hjjang.backend.domain.image.domain.entity.Image;
+import com.hjjang.backend.domain.image.exception.ImageNotFoundException;
+import com.hjjang.backend.infra.aws.service.S3ImageUploader;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.transaction.Transactional;
 import java.io.IOException;
 import java.util.List;
 
-public interface ImageService {
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class ImageService {
+    private final ImageRepository imageRepository;
 
-    List<Image> findAll();
+    private final S3ImageUploader imageUploader;
 
-    Image findImageById(Long imageId);
+    public List<Image> findAll() {
+        return imageRepository.findAll();
+    }
 
-    Image uploadNewImage(MultipartFile multipartFile) throws IOException;
+    public Image findImageById(Long imageId) {
+        return imageRepository.findById(imageId).orElseThrow(ImageNotFoundException::new);
+    }
 
-    void removeImage(Image image);
+    public Image uploadNewImage(MultipartFile multipartFile) throws IOException {
+        return imageRepository.save(new Image(imageUploader.upload(multipartFile)));
+    }
+
+    public void removeImage(Image image) {
+        image.removeImage();
+        imageRepository.save(image);
+    }
 
 }
