+++ backend/src/main/java/com/hjjang/backend/domain/image/service/LocalImageUploader.java
@@ -1,0 +1,10 @@
+package com.hjjang.backend.domain.image.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@RequiredArgsConstructor
+@Service
+public class LocalImageUploader extends ImageUploader{}
