+++ backend/src/main/java/com/hjjang/backend/infra/aws/service/S3ImageUploader.java
@@ -1,24 +1,22 @@
-package com.hjjang.backend.infra.uploader.component;
+package com.hjjang.backend.infra.aws.service;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.hjjang.backend.domain.image.service.ImageUploader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Optional;
-import java.util.UUID;
 
 @Slf4j
 @RequiredArgsConstructor
-@Component
-public class S3Uploader implements ImageUploader {
+@Service
+public class S3ImageUploader extends ImageUploader {
 
     private final AmazonS3Client amazonS3Client;
 
@@ -33,11 +31,6 @@ public class S3Uploader implements ImageUploader {
                                 () -> new IllegalArgumentException("error: MultipartFile -> File convert fail"));
 
         return upload(uploadFile);
-    }
-
-    @Override
-    public String changeFileName(MultipartFile uploadFile) {
-        return UUID.randomUUID() + ".png"; // 저장될 파일 이름 변환
     }
 
     // S3로 파일 업로드하기
@@ -62,20 +55,5 @@ public class S3Uploader implements ImageUploader {
             return;
         }
         log.info("File delete fail");
-    }
-
-    // 로컬에 파일 업로드 하기
-    @Override
-    public Optional<File> localUpload(MultipartFile file) throws IOException {
-        File convertFile = new File(Path.imageSavePath.getPath() + changeFileName(file));
-        if (convertFile.createNewFile()) { // 바로 위에서 지정한 경로에 File이 생성됨 (경로가 잘못되었다면 생성 불가능)
-            try (FileOutputStream fos =
-                         new FileOutputStream(convertFile)) { // FileOutputStream 데이터를 파일에 바이트 스트림으로 저장하기 위함
-                fos.write(file.getBytes());
-            }
-            return Optional.of(convertFile);
-        }
-
-        return Optional.empty();
     }
 }
